Add debuging tools + Solve Bug on robot update
Robot update was called more than one time by loops.

+++ src/ai/math/ContinuousAngle.cpp
@@ -95,7 +95,7 @@ ContinuousAngle& ContinuousAngle::operator*=( double scalar ){
     return *this;
 }
 
-float ContinuousAngle::turn() const {
+double ContinuousAngle::turn() const {
     return this->angle_value /(2*M_PI);
 };
